# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Service for processing orders at Pick-Up Point. Allow registration and login by email/password as well as dummy login using user roles (client, employee, moderator).",
        "title": "Order Pick-Up Point",
        "contact": {
            "name": "Egor Ponyaev",
            "url": "https://github.com/epchamp001",
            "email": "epchamp001@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/dummyLogin": {
            "post": {
                "description": "Get a JWT token by passing a desired user role (client, employee, moderator) through dummy login.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Dummy login for testing",
                "parameters": [
                    {
                        "description": "Dummy login request with role",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DummyLoginPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "$ref": "#/definitions/dto.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid role or error during dummy login",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Login a user using email and password. Returns a JWT token if credentials are valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JWT token",
                        "schema": {
                            "$ref": "#/definitions/dto.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized: invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    }
                }
            }
        },
...
        "/register": {
            "post": {
                "description": "Register a new user with email, password and role (client, employee, moderator).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterPostRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registration success response with user ID",
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.DummyLoginPostRequest": {
            "description": "Request payload for dummy login. Provide a desired user role (\"client\", \"employee\", \"moderator\") to obtain a JWT token.",
            "type": "object",
            "properties": {
                "role": {
                    "type": "string",
                    "example": "employee"
                }
            }
        },
        "dto.Error": {
            "description": "Standard error response containing an error message.",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "dto.LoginPostRequest": {
            "description": "Request payload for login using email and password.",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "strongpassword123"
                }
            }
        },
...
        "dto.RegisterPostRequest": {
            "description": "Request payload for user registration.",
            "type": "object",
            "required": [
                "email",
                "password",
                "role"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "strongpassword123"
                },
                "role": {
                    "type": "string",
                    "example": "moderator"
                }
            }
        },
        "dto.RegisterResponse": {
            "description": "Response returned after successful user registration.",
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "example": "user123"
                }
            }
        },
        "dto.TokenResponse": {
            "description": "Response containing a JWT token.",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token. Obtain the token via /login (using email and password) or via /dummy Login (passing desired role).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}